clip_baseline/requirements.txt
torch
torchvision
transformers
Pillow
numpy
scikit-learn
tqdm
openai-clip
ftfy
regex
